whatsapp-supabase-backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
python-dotenv==1.0.0
pydantic==2.4.2
supabase==2.0.0
python-multipart==0.0.6
selenium==4.15.2
webdriver-manager==4.0.1
loguru==0.7.2
python-jose==3.3.0
passlib==1.7.4
httpx
bcrypt==4.0.1
pytest==7.4.3
supabase
email-validator
